Remove extra crates (#6)
