+++ Cargo.toml
@@ -1,23 +1,22 @@
 [package]
-name = "rsnode"
+name = "rs-node"
 version = "0.1.0"
 edition = "2021"
 autobenches = false
 
 [workspace]
 members = [
-    "cli",
-    "client",
-    "common",
-    "derivation",
     "backend",
-    "execution",
 ]
 
 [dependencies]
-cli = { path = "./cli" }
-client = { path = "./client" }
-common = { path = "./common" }
-derivation = { path = "./derivation" }
 backend = { path = "./backend" }
-execution = { path = "./execution" }
+ethers-core = "1.0.2"
+ethers-providers = "1.0.2"
+futures = "0.3.26"
+serde = { version = "1.0.152", features = ["derive"] }
+serde_json = "1.0.92"
+tokio = { version = "1.25.0", features = ["macros", "rt-multi-thread"] }
+eyre = "0.6.8"
+dotenv = "0.15.0"
+
