--- cli/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "cli"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ethers-core = "1.0.2"
-ethers-providers = "1.0.2"
-serde = { version = "1.0.152", features = ["derive"] }
-serde_json = "1.0.92"
-eyre = "0.6.8"
-dotenv = "0.15.0"
-
-client = { path = "../client" }
