--- common/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "common"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ethers-core = "1.0.2"
-serde = { version = "1.0.152", features = ["derive"] }
-serde_json = "1.0.92"
